+++ egui-examples/src/lib.rs
@@ -15,7 +15,7 @@ pub struct MyApp;
 
 impl App for MyApp {
     fn ui(&mut self, ui: &mut Ui, _frame: &mut Frame) {
-        CentralPanel::default().show_inside(ui, |ui| {
+        CentralPanel::default().show(ui, |ui| {
             ui.heading("My egui Android App");
             ui.label("Hello world!");
         });